requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.36
psycopg[binary]>=3.2.10
python-dotenv>=1.0.1
bcrypt>=4.0.0
python-jose[cryptography]>=3.3.0
requests>=2.31.0
pydantic-settings>=2.2.1
alembic>=1.14.0
